+++ barretenberg/cpp/src/barretenberg/eccvm/eccvm_composer.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
-#include "barretenberg/honk/proof_system/eccvm_prover.hpp"
-#include "barretenberg/honk/proof_system/eccvm_verifier.hpp"
+#include "barretenberg/eccvm/eccvm_prover.hpp"
+#include "barretenberg/eccvm/eccvm_verifier.hpp"
 #include "barretenberg/proof_system/circuit_builder/eccvm/eccvm_circuit_builder.hpp"
 #include "barretenberg/proof_system/composer/composer_lib.hpp"
 #include "barretenberg/srs/factories/file_crs_factory.hpp"
