+++ barretenberg/cpp/src/barretenberg/eccvm/eccvm_composer.test.cpp
@@ -3,7 +3,7 @@
 #include <gtest/gtest.h>
 #include <vector>
 
-#include "barretenberg/honk/composer/eccvm_composer.hpp"
+#include "barretenberg/eccvm/eccvm_composer.hpp"
 #include "barretenberg/numeric/uint256/uint256.hpp"
 #include "barretenberg/polynomials/polynomial.hpp"
 #include "barretenberg/proof_system/circuit_builder/eccvm/eccvm_circuit_builder.hpp"
