+++ barretenberg/cpp/src/barretenberg/eccvm/eccvm_prover.cpp
