+++ barretenberg/cpp/src/barretenberg/eccvm/eccvm_verifier.hpp
